Coding Days: grid generators
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2751 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/generation/TruncatedTrapezoidGridGenerator.cpp
@@ -43,7 +43,7 @@ void TruncatedTrapezoidGridGenerator::truncated(int level,size_t l_user)
 	   throw generation_exception("Grid level value is negative");
 	}
 	HashGenerator gen;
-	gen.truncated(this->storage, static_cast<HashGenerator::level_t>(level),l_user);
+	gen.truncated(this->storage, static_cast<HashGenerator::level_t>(level), static_cast<int>(l_user));
 }
 
 //void BoundaryGridGenerator::refine(RefinementFunctor* func)
